stockfish.wasm: Do not create worker for fake thread

+++ src/thread.h
@@ -44,12 +44,12 @@ class Thread {
 
   Mutex mutex;
   ConditionVariable cv;
-  size_t idx;
+  int idx;
   bool exit = false, searching = true; // Set before starting std::thread
   pthread_t nativeThread;
 
 public:
-  explicit Thread(size_t);
+  explicit Thread(int);
   virtual ~Thread();
   virtual void search();
   void clear();
